+++ packages/profile-sync-controller/src/controllers/authentication/index.ts
@@ -4,6 +4,7 @@ export { AuthenticationController as Controller };
 export default AuthenticationController;
 export * from './AuthenticationController.js';
 export * as Mocks from './mocks/index.js';
+export { authenticationControllerSelectors } from './selectors.js';
 
 export type {
   AuthenticationControllerPerformSignInAction,